Add custom hook 'useAuthCheck'
